third_party/blink/web_tests/svg/dynamic-updates/SVGEllipseElement-svgdom-cx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cx' property of the SVGEllipseElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ellipseElement.cx.baseVal.value is -50
PASS ellipseElement.cx.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
